Adjust remaining tests and router placement
Include docs links in tiles, where applicable.
