+++ lib/plausible_web/controllers/settings_controller.ex
@@ -51,6 +51,40 @@ defmodule PlausibleWeb.SettingsController do
     render(conn, :api_keys, layout: {PlausibleWeb.LayoutView, :settings}, api_keys: api_keys)
   end
 
+  def new_api_key(conn, _params) do
+    changeset = Auth.ApiKey.changeset(%Auth.ApiKey{})
+
+    render(conn, "new_api_key.html", changeset: changeset)
+  end
+
+  def create_api_key(conn, %{"api_key" => %{"name" => name, "key" => key}}) do
+    case Auth.create_api_key(conn.assigns.current_user, name, key) do
+      {:ok, _api_key} ->
+        conn
+        |> put_flash(:success, "API key created successfully")
+        |> redirect(to: Routes.settings_path(conn, :api_keys) <> "#api-keys")
+
+      {:error, changeset} ->
+        render(conn, "new_api_key.html", changeset: changeset)
+    end
+  end
+
+  def delete_api_key(conn, %{"id" => id}) do
+    case Auth.delete_api_key(conn.assigns.current_user, id) do
+      :ok ->
+        conn
+        |> put_flash(:success, "API key revoked successfully")
+        |> redirect(to: Routes.settings_path(conn, :api_keys) <> "#api-keys")
+
+      {:error, :not_found} ->
+        conn
+        |> put_flash(:error, "Could not find API Key to delete")
+        |> redirect(to: Routes.settings_path(conn, :api_keys) <> "#api-keys")
+    end
+  end
+
+
+
   def danger_zone(conn, _params) do
     render(conn, :danger_zone, layout: {PlausibleWeb.LayoutView, :settings})
   end
@@ -129,7 +163,6 @@ defmodule PlausibleWeb.SettingsController do
   end
 
   def cancel_update_email(conn, _params) do
-    IO.inspect :HIT
     changeset = Auth.User.cancel_email_changeset(conn.assigns.current_user)
 
     case Repo.update(changeset) do
@@ -198,6 +231,16 @@ defmodule PlausibleWeb.SettingsController do
       password_changeset: password_changeset,
       layout: {PlausibleWeb.LayoutView, :settings}
     )
+  end
+
+  def delete_session(conn, %{"id" => session_id}) do
+    current_user = conn.assigns.current_user
+
+    :ok = UserAuth.revoke_user_session(current_user, session_id)
+
+    conn
+    |> put_flash(:success, "Session logged out successfully")
+    |> redirect(to: Routes.settings_path(conn, :security) <> "#user-sessions")
   end
 
   defp do_update_password(user, params) do
