+++ lib/plausible_web/templates/auth/generate_2fa_recovery_codes.html.heex
@@ -40,7 +40,7 @@
     <button
       id="finish"
       class="button w-full mt-4"
-      onclick={"location.replace('#{Routes.auth_path(@conn, :user_settings) <> "#setup-2fa"}')"}
+      onclick={"location.replace('#{Routes.settings_path(@conn, :security) <> "#update-2fa"}')"}
     >
       Finish
     </button>
