+++ lib/plausible_web/templates/auth/initiate_2fa_setup.html.heex
@@ -11,7 +11,7 @@
     <.focus_list>
       <:item>
         Changed your mind?
-        <.styled_link href={Routes.auth_path(@conn, :user_settings) <> "#setup-2fa"}>
+        <.styled_link href={Routes.settings_path(@conn, :security) <> "#update-2fa"}>
           Go back to Settings
         </.styled_link>
       </:item>
