+++ lib/plausible_web/templates/auth/user_settings.html.heex
@@ -456,80 +456,6 @@
     </div>
   </div>
 
-  <div
-    id="api-keys"
-    class="max-w-2xl px-8 pt-6 pb-8 mx-auto mt-16 bg-white border-t-2 border-indigo-100 rounded rounded-t-none shadow-md dark:bg-gray-800 dark:border-indigo-500"
-  >
-    <h2 class="text-xl font-black dark:text-gray-100">API Keys</h2>
-    <div class="my-4 border-b border-gray-300 dark:border-gray-500"></div>
-
-    <PlausibleWeb.Components.Billing.Notice.premium_feature
-      billable_user={@current_user}
-      current_user={@current_user}
-      feature_mod={Plausible.Billing.Feature.StatsAPI}
-    />
-
-    <div class="flex flex-col mt-6">
-      <div class="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
-        <div class="inline-block min-w-full py-2 align-middle sm:px-6 lg:px-8">
-          <%= if Enum.any?(@api_keys) do %>
-            <div class="overflow-hidden border-b border-gray-200 shadow dark:border-gray-900 sm:rounded-lg">
-              <table class="min-w-full divide-y divide-gray-200 dark:divide-gray-900">
-                <thead class="bg-gray-50 dark:bg-gray-900">
-                  <tr>
-                    <th
-                      scope="col"
-                      class="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase dark:text-gray-100"
-                    >
-                      Name
-                    </th>
-                    <th
-                      scope="col"
-                      class="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase dark:text-gray-100"
-                    >
-                      Key
-                    </th>
-                    <th scope="col" class="relative px-6 py-3">
-                      <span class="sr-only">Revoke</span>
-                    </th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <%= for api_key <- @api_keys do %>
-                    <tr class="bg-white dark:bg-gray-800">
-                      <td class="px-6 py-4 text-sm font-medium text-gray-900 dark:text-gray-100 whitespace-nowrap">
-                        <%= api_key.name %>
-                      </td>
-                      <td class="px-6 py-4 text-sm text-gray-500 dark:text-gray-100 whitespace-nowrap">
-                        <%= api_key.key_prefix %><%= String.duplicate("*", 32 - 6) %>
-                      </td>
-                      <td class="px-6 py-4 text-sm font-medium text-right whitespace-nowrap">
-                        <%= button("Revoke",
-                          to: "/settings/api-keys/#{api_key.id}",
-                          class: "text-red-600 hover:text-red-400 dark:text-red-500",
-                          method: :delete,
-                          "data-confirm": ""
-                        ) %>
-                      </td>
-                    </tr>
-                  <% end %>
-                </tbody>
-              </table>
-            </div>
-          <% end %>
-
-          <PlausibleWeb.Components.Generic.button_link
-            :if={Plausible.Billing.Feature.StatsAPI.check_availability(@current_user) == :ok}
-            href={Routes.auth_path(@conn, :new_api_key)}
-            class="mt-4"
-          >
-            + New API Key
-          </PlausibleWeb.Components.Generic.button_link>
-        </div>
-      </div>
-    </div>
-  </div>
-
   <div class="max-w-2xl px-8 pt-6 pb-8 mx-auto mt-16 mb-24 bg-white border-t-2 border-red-600 rounded rounded-t-none shadow-md dark:bg-gray-800">
     <div class="flex items-center justify-between">
       <h2 class="text-xl font-black dark:text-gray-100">Delete account</h2>
