+++ lib/plausible_web/templates/settings/danger_zone.html.heex
@@ -3,7 +3,7 @@
 </.notice>
 
 <.settings_tiles>
-  <.tile>
+  <.tile docs="delete-account">
     <:title>Delete Account</:title>
     <:subtitle>Deleting your account removes all sites and stats you've collected</:subtitle>
