+++ lib/plausible_web/templates/settings/preferences.html.heex
@@ -20,12 +20,12 @@
     </.form>
   </.tile>
 
-  <.tile>
+  <.tile docs="dashboard-appearance">
     <:title>
       <a id="update-theme">Dashboard Appearance</a>
     </:title>
     <:subtitle>
-      Set your ... preferences
+      Set your visual preferences
     </:subtitle>
     <.form
       :let={f}
