+++ lib/plausible_web/templates/settings/subscription.html.heex
@@ -1,5 +1,5 @@
 <.settings_tiles>
-  <.tile>
+  <.tile docs="billing">
     <:title>
       <a id="subscription">Subscription</a>
     </:title>
