+++ lib/plausible_web/templates/stats/site_locked.html.heex
@@ -36,7 +36,7 @@
             </p>
           </div>
           <div class="mt-6 w-full text-center">
-            <.button_link href={Routes.auth_path(@conn, :user_settings)}>
+            <.button_link href={Routes.settings_path(@conn, :subscription)}>
               Manage my subscription
             </.button_link>
           </div>
